# public/swagger-specs/services/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ConsenSys Codefi Orchestrate API. Enables dynamic management of transactions, identities, chains, faucets and contracts.\nTransaction Requests are an abstraction over schedules and jobs representing one or more transactions executed on the Blockchain network\nSchedules are ordered lists of jobs executed in a predefined sequence\nJobs represent single transaction flows executed on the Blockchain network\nChains represent list of endpoints pointing to a Blockchain network\nFaucets represent funded accounts (holding ETH) linked to specific chains, allowed to fund newly created accounts automatically for them to be able to send transactions.\nAccounts represent Ethereum accounts (private keys). By usage of the generated cryptographic key pair, accounts can be used to sign/verify and to encrypt/decrypt messages.\nContracts represent Solidity contracts management.",
        "title": "Orchestrate API",
        "contact": {
            "name": "Contact ConsenSys Codefi Orchestrate",
            "url": "https://consensys.net/codefi/orchestrate/contact",
            "email": "orchestrate@consensys.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get a list of filtered transaction requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Search transaction requests by provided filters",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "List of idempotency keys",
                        "name": "idempotency_keys",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transaction requests found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.TransactionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filter in the request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/deploy-contract": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new contract deployment request\nThe transaction can be private (Tessera, EEA).\nThe transaction can be a One Time Key transaction in 0 gas private networks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Creates and sends a new contract deployment",
                "parameters": [
                    {
                        "description": "Deployment transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.DeployContractRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/api.DeployContractParams"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "gasPricePolicy": {
                                                            "allOf": [
                                                                {
                                                                    "$ref": "#/definitions/api.GasPriceParams"
                                                                },
                                                                {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "retryPolicy": {
                                                                            "$ref": "#/definitions/api.RetryParams"
                                                                        }
                                                                    }
                                                                }
                                                            ]
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created deployment transaction request",
                        "schema": {
                            "$ref": "#/definitions/api.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/send": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new smart contract transaction request\nThe transaction can be private (Tessera, EEA).\nThe transaction can be a One Time Key transaction in 0 gas private networks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Creates and sends a new contract transaction",
                "parameters": [
                    {
                        "description": "Contract transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SendTransactionRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/api.TransactionParams"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "gasPricePolicy": {
                                                            "allOf": [
                                                                {
                                                                    "$ref": "#/definitions/api.GasPriceParams"
                                                                },
                                                                {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "retryPolicy": {
                                                                            "$ref": "#/definitions/api.RetryParams"
                                                                        }
                                                                    }
                                                                }
                                                            ]
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created contract transaction request",
                        "schema": {
                            "$ref": "#/definitions/api.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/send-raw": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new raw transaction request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Creates and sends a raw transaction",
                "parameters": [
                    {
                        "description": "Raw transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.RawTransactionRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/api.RawTransactionParams"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "retryPolicy": {
                                                            "$ref": "#/definitions/api.IntervalRetryParams"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created raw transaction request",
                        "schema": {
                            "$ref": "#/definitions/api.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/transactions/{uuid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Fetch a single transaction request by uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Fetch a transaction request by uuid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of the transaction request",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction request found",
                        "schema": {
                            "$ref": "#/definitions/api.TransactionResponse"
                        }
                    },
                    "404": {
                        "description": "Transaction request not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.GasPriceParams": {
            "type": "object",
            "properties": {
                "priority": {
                    "type": "string",
                    "example": "very-high"
                },
                "retryPolicy": {
                    "$ref": "#/definitions/api.RetryParams"
                }
            }
        },
...
        "httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 24000
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}